fix: user code refactoring
