+++ src/main/java/com/t1f5/skib/global/enums/UserType.java
@@ -1,4 +1,4 @@
-package com.t1f5.skib.user.model;
+package com.t1f5.skib.global.enums;
 
 public enum UserType {
     TRAINER,
