+++ src/main/java/com/t1f5/skib/user/dto/requestdto/RequestUpdateUserDto.java
@@ -2,12 +2,14 @@
 package com.t1f5.skib.user.dto.requestdto;
 
 import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
-@Getter
-@Setter
 
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
 public class RequestUpdateUserDto {
     private Integer userId;
