+++ src/main/resources/application.properties
@@ -1,9 +1,9 @@
 spring.application.name=skib
 
 # mariaDB 설정
-spring.datasource.url=jdbc:mariadb://localhost:3306/SKIB
+spring.datasource.url=jdbc:mariadb://localhost:3306/skib?useSSL=false&disableGssapi=true
 spring.datasource.username=root
-spring.datasource.password=1234
+spring.datasource.password=0000
 spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
 
 # Hibernate 자동 설정 옵션
